+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## unreleased
 * Breaking Changes
+  * CardPayments
+    * Remove `status` property from `CardResult`
   * CorePayments
     * CoreConfig instances must now be instantiated using a `clientId` instead of an `accessToken`